API Changes: Sol-str-slice
Replace the caseeq keyword withcase_eq

sol_str_slice_to_int() now receive a pointer to long int.

issue #1663

+++ src/lib/datatypes/sol-str-slice.c
@@ -25,7 +25,7 @@
 #include "sol-util-internal.h"
 
 SOL_API int
-sol_str_slice_to_int(const struct sol_str_slice s, int *value)
+sol_str_slice_to_int(const struct sol_str_slice s, long int *value)
 {
     char *endptr = NULL;
     long int v;
